[TASK] File-List should allow quick access to translations
File-Metadata has been configured to be translatable.
In List-Module we are used to also see the translated
records. To help editors this possibility is introduced
to the file list to.

Based on the fact that for a file all listed properties
would be the same than in every language, a different
implementation has been used.

The icons and the formatting is prelimary and subject
to be changed by the usability team.

Resolves: #52817
Releases: 6.2
Change-Id: I0016b8c9fe543d3bab3e77dc37ae7bb67acc97b0
Reviewed-on: https://review.typo3.org/24752
Reviewed-by: Ernesto Baschny
Tested-by: Ernesto Baschny
Reviewed-by: Wouter Wolters
Reviewed-by: Fabien Udriot
Tested-by: Fabien Udriot
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers