[TASK] Improve usability of Filelist "Drag-Upload"
We now can drag & drop files in the file list to upload.
But it looks ugly and has some issues.

This change adds a progress overview as provided by the
UX team and improves the usability.
The progress is shown for files added by drag & drop but
also when you use the file picker.

We switched to the jQuery ajax function as this has
a generic event and response handler.

Resolves: #51361, #53829
Release: 6.2
Change-Id: Ifd7b189f84594149b9590bbd4058ad68e4439344
Reviewed-on: https://review.typo3.org/27064
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Stefan Froemken
Tested-by: Stefan Froemken
Reviewed-by: Benjamin Mack
Tested-by: Benjamin Mack