[BUGFIX] Do not show exeption when folder is not accesible
When a storage folder is not accessible for the user a exception
is displayed. Same if the folder is missing.

This patch changes the exception to
\TYPO3\CMS\Core\Resource\Exception.

The user will get a flash message and FAL tries to display the
parent folder of the restricted folder. If the folder is
missing the parent folder is displayed.

Resolves: #51762, #51319
Releases: 6.2
Change-Id: I1fb930f8c13049eeddb27521326f1d0fc8707933
Reviewed-on: https://review.typo3.org/27189
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter