[TASK] BasicFileUtility::init is not deprecated yet
Until we fully remove usage of BasicFileUtility throughout the core
(i.e. DataHandler, Import/Export module, ElementBrowser and others),
the core should not pollute the deprecation log with it.

We just individually deprecate methods which are not used anymore.
Methods which are still used in the core are marked as @deprecated
(to visualize this in the IDE) but they won't write to the
deprecation log yet.

Resolves: #57209
Releases: 6.2
Change-Id: I41d7f6c8b7c3c4df77791ce3e6f3f69140d1935d
Reviewed-on: https://review.typo3.org/28674
Reviewed-by: Xavier Perseguers
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters