[FEATURE] Remove conf.php for ext:filelist
In order to be able to remove the conf.php a few fixes are needed.
The module dispatcher checks for a conf.php file and results in a error.
The language labels for the module menu for example are missing currently.
This patch adds a possibility to register the labels also in addModule
function.

Resolves: #62880
Releases: master
Change-Id: I5765cf84858d1c328739c761d23c97307abf7ede
Reviewed-on: http://review.typo3.org/33992
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>
Reviewed-by: Steffen Müller <typo3@t3node.com>
Tested-by: Steffen Müller <typo3@t3node.com>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>