[BUGFIX] sys_file uid should be visible in filelist
There are several cases where editors manually
link to files (<LINK file:123>Download PDF</LINK>),
f.e. in bulletlists, so they need a easy way to
find the uid of a specific file. A good place would
be the mouseover-text of files in the filelist –
like as the pages or tt_content elements have their
uid shown in the element's icon mouseover text.

Resolves: #59866
Releases: master, 6.2
Change-Id: Iec3e5043d01bfb3fac60d738ef7749fc0126c6b2
Reviewed-on: http://review.typo3.org/35731
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>