[BUGFIX] Properly check type of POST vars on clipboard action
Avoid a PHP warning by checking the type of data retrieved via POST.

"CBC" is an array of checkboxes, which can be null if nothing is selected.
"CBH" is an array of hidden fields. If this is not an array, something
is messed up in HTML generation and an Exception should be thrown then
(because of a PHP warning).

Also adapt the list module code accordingly.

Resolves: #64897
Releases: master, 6.2
Change-Id: Ife7d079a6d666e8bdbbce7b7eef34dbf5636c425
Reviewed-on: http://review.typo3.org/36700
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>