[BUGFIX] Make editIconsHook work again
It isn't possible to add icons in filelist extended view easily,
because there is no way to get the current file or folder object nor get
information about it. This patch adds the object to the cells variable
and reset it after hook execution to ensure access to it.

Releases: master, 6.2
Resolves: #67962
Change-Id: I44a4229128bcfdb2ed17900462f50a771ed63fdd
Reviewed-on: http://review.typo3.org/41117
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>