[TASK] Render Resources using SVG in IconFactory
Add SVG icons for MimeTypes to the IconRegistry, add
IconFactory->getIconForResource() as improved version of
IconUtility->getSpriteIconForResource().

Also add IconFactory->getIconForFileExtension() similar to
IconUtility->getSpriteIconForFile() but without the possibility to pass
a path but only a file extension.

Replace all core calls of getSpriteIconForFile() and
getSpriteIconForResource() with calls to IconFactory.

Resolves: #69744
Resolves: #69759
Releases: master
Change-Id: I26b221fbba87c7f3e09ec10b61c13c8e74b8ecb3
Reviewed-on: http://review.typo3.org/43258
Reviewed-by: Benjamin Kott <info@bk2k.info>
Tested-by: Benjamin Kott <info@bk2k.info>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>