[TASK] Fix CGL violation in FileList
https://review.typo3.org/c/Packages/TYPO3.CMS/+/57414 introduced
a CGL violation that was not discovered because the patchset was
pushed on 8th of august and was checked with php-cs-fixer 2.14.1
on bamboo where no violation was found.

In the meantime (between push and merge of the mentioned patch)
https://review.typo3.org/c/Packages/TYPO3.CMS/+/61564 has been
merged, which updates php-cs-fixer to 2.15.2 and contains new
rules.

Releases: master
Resolves: #89282
Related: #57414
Related: #85425
Change-Id: I61d5e6e5dc61b500fe620e59002f296f50f03214
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61835
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Björn Jacob <bjoern.jacob@tritum.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Björn Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>