[TASK] Simplify code of FileList module
The FileList class was originally a derivative of DatabaseRecordList,
which was decoupled in TYPO3 v9 and and marked as internal.

It is responsible for showing the actual "FileList" in the TYPO3 Backend
module.

A lot of this code has grown historically, and can be removed and
simplified, as it is not directly connected to the abstract record list anymore.

This patch removes unused public properties like $JScode as this is
part of a "multi-table" setup as is possible in the database listings.

Some methods are split into smaller parts to make the methods
themselves smaller.

This is however just the first step, next to separating the curation
and the output (and moving this into a fluid template), also using proper pagination
for fetching files and folders, or building DI for the class now is much easier to do.

Resolves: #90242
Releases: master
Change-Id: Ie66964d9df355a9d74dbe9d0a8a17f9cecdf7c83
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63067
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Susanne Moog <look@susi.dev>