[TASK] Extbase controller actions must return ResponseInterface
In the past, Extbase controller actions could return either void,
null, a string, or an object that implemented the __toString()
method. All those return types are now deprecated in favor of
an instance of \Psr\Http\Message\ResponseInterface (PSR-7).

Response objects can be created in the controller by using the
response factory which is injected into all action controllers.

Alternatively, custom response objects can be declared and used
as long as they implement the interface and are therefore PSR-7
compatible.

Releases: master
Resolves: #92784
Change-Id: If1e778c29f870fd0e78d253f0cfcff359a0babae
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66553
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Daniel Haupt <mail@danielhaupt.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Daniel Haupt <mail@danielhaupt.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>