[TASK] Process _PATH_ field also in directory listing
In #87974 the special _PATH_ field was added to
FileList, to be displayed on the search result view.

To be consistent, this special field is now also
processed on building the directory listing, while
it currently has no visual effect, since the _PATH_
field is only dynamically added to in a search
context.

Resolves: #94505
Releases: master
Change-Id: I100ab601dcf28f7acdb9bf999cf08cfa350ffd4f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69760
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>