[TASK] Streamline clipboard UI
The menu was removed because it contained
only already existing buttons (clear clipboard)
or buttons that do not belong to the clipboard
(delete, export) which made the clipboard
overly complicated.

Removed red color (danger) from clear all
button, because this is a non destructive
operation and added width/height for preview
images so SVG images are displayed as expected
in the clipboard.

In addition, the "move/copy" selector is now a
radio button, which finally works properly again.

Resolves: #94507
Releases: master
Change-Id: Ie9d98b13fa3075f6f8d297754576e3c20fd69b85
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69798
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>