[TASK] Move metadata translation to action button in filelist
Creating and editing file metadata translations is usually
done in the filelist module. Therefore a custom toggle
button in the "Localization" column had to be used.

In the expanded state, various flag icons with an hard
to guess overlay icon were displayed. Those could be
used to either create or edit the file metadata records.

This however had some drawbacks:

- The toggle button was not accessible by keyboard
- As toggle icon, the "pages languages overlay" icon
  was used, making it hard to understand for editors
- As soon as one of those toggle buttons was clicked,
  all other toggle buttons jumped to the left

To improve this, the toggle button is moved into the
controls group and does now uses the "actions-translate"
icon. It furthermore does now open a dropdown with
proper link labels - as already known from the secondary
menu. This therefore also increases consistency.

Besides the visual changes, this patch also cleans
up the underlying code and removes the custom
JavaScript component, which was used for the
toggle button.

Resolves: #94997
Releases: master
Change-Id: I898e234d35b125ba37ac6c30874fa206d2515184
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70759
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>