[BUGFIX] Fix clipboard handling in filelist
This resolves some problems on using the clipboard
in the filelist module. Previously it was for
example not possible to transfer files / folders
to the clipboard again, after it was cleared,
since the corresponding GET parameter was kept
in the URL. This is now fixed by assigning the
action to the filelist form.

Further topics:

- Pagination does now use the correct colspan
- setCB and delete actions are now handled
  via multiple record selection events and do
  not longer rely on "trigger classes"
- the superfluous internal "selectedElements"
  array is removed again, since the display of
  actions is handled by the multi record selection
  component anyways

Resolves: #95055
Releases: master
Change-Id: Iff15716e09d149ed29d72b1ae45e88ba2a43b92d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70841
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>