[TASK] Add missing configuration options for file download
With #90548 the filelist module has been extended
for the possibility to download files and folders.

Since file download might not be enabled for all
users, or should at least be limited to a set of
file extensions, this patch adds the necessary
configuration options.

It's therefore now possible to either specify
allowed / disallowed file extensions or to
completely disable the file download.

In case the download can not be prepared due
to no valid files were found, the user
receives a notification.

Resolves: #95113
Related: #90548
Releases: master
Change-Id: Ifc4b4a5cd4ac8cb49b14e8eb6239ece37596e7e4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71013
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>