[BUGFIX] Avoid tooltip for secondary options in list modules
The secondary options in list modules (recordlist,
filelist) are displayed in a dropdown with an icon and
the corresponding title. Therefore, an additional title
attribute is not necessary, but actually leads to a tooltip
being triggered with the same information again.

This is fixed by removing the title attribute in case the
title was already added as button tag content.

Resolves: #95462
Releases: master
Change-Id: I520790d7d0eb9283543c50a5c62f71c5f60918ca
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71426
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>