[TASK] Require moduleData to be set in request
Since #96895, the ModuleData object is created
for every request, targeting a backend module, and
attached to the PSR-7 Request object.

Previously the route targets (the controllers) made
use of the nullsafe operator when accessing the
module data. This however hides the dependency to
the object. In case the module data is missing, the
controller should fail "hard".

Therefore, those nullsafe operators are removed again.

Additionally, the $default value is removed from the
moduleData->get() calls, since the requested properties
are all registered in the module registration with a
corresponding default value.

Resolves: #96917
Related: #96895
Releases: main
Change-Id: I3d53e435a809e925f18c7495d832beb94c7f735b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73554
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>