[BUGFIX] Fix exception in filelist when nullable columns are shown
When nullable columns are shown in filelist module, an
exception is thrown, because null may be passed to
htmlspecialchars().

This change ensures that an empty string is passed to
htmlspecialchars(), if the column value is null. Also
the last parameter from the getProcessedValueExtra()
function call is removed, since it equals the functions
default value for the $forceResult parameter.

Resolves: #97316
Releases: main, 11.5
Change-Id: I7d82aea0ed073f6f30944d53e0803ce848e30e53
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74145
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>