[BUGFIX] Handle sorting of paginated FileList properly
Since #97443 it's finally possible to sort by
all the metadata fields, available since the
introduction of the column selector in #96959.

The changes in #97443 however lead to the
fact that pagination and sorting did no longer
work together because pagination was still
applied by the driver, while sorting was done
afterwards.

This is now finally resolved. In case sorting
is requested, both the pagination and sorting
is applied manually after fetching files and
folders from the driver in the correct order.

Resolves: #97467
Related: #97443
Related: #96959
Releases: main, 11.5
Change-Id: Ie9363d13541be20ef1b118fa1d627486d6ce929e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74395
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>