[BUGFIX] Cast `$chars` argument of `fixed_lgd_cs()` to int
Since #97578, `GeneralUtility` is declared strict, potentially breaking
calls to `fixed_lgd_cs()` in at least the "List" module. The argument
`$chars` must be of type int, but sometimes it's a string which breaks
the call of `abs()`.

All calls to this method are adjusted, where required, and the value is
explicitly cast to int.

Resolves: #97887
Related: #97578
Releases: main
Change-Id: I7feed1ae8dbe7cdbc906d73cfbc49f7d9dbeb6fe
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75091
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>