[TASK] Pass dataset to context menu actions and drop jQuery in them
The context menu module now passes the dataset bound to a context menu
item to the respective callback action, excluding `callback-action` and
`callback-module`.

The context menu items bound to `this` in the respective callback actions
are not required anymore, allowing to ditch jQuery in the
`context-menu-actions` modules. For this reason, binding the context
menu item to `this` is deprecated, usages will trigger a deprecation log
entry in the browser's console.

Resolves: #98168
Releases: main
Change-Id: Ibaf3c5b343323dcd29bd9aae3212ebcee91d83ef
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75491
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>