[TASK] Separate file + folder creation controllers
This change splits up the functionality of
the "CreateFolderController" into "add folder"
and "add file".

This also affects:
* Buttons in the file list module
* Actions in the Context Menu

which also uses more specific labels + icons.

This way, it is much clearer for editors on specific
actions. In addition, the code is split up
into CreateFileController and CreateFolderController.

Separate steps could be to also split up
the online media functionality into a modal.

Resolves: #98999
Releases: main
Change-Id: I7ccbae09de6b9fa39996e0357129006091601e27
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76433
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Bartsch <bo@cedev.de>