[FEATURE] Add tile view to filelist
The listing of resources in a table has some specific use
cases but for the editor, hard to get an overview of file
resources since the thumbnails are small.

To provide a better overview of assets in the file module we
are introducing a tile view with bigger thumbnails and reduced
meta information. The user can now choose the desired display
mode of the assets depending on the current requirements.

The user can change the view mode in the view menu in the
module menu bar. By default the tile view is enabled for
new and existing users.

TYPO3 will remember the choice of the user.

Resolves: #99155
Releases: main
Change-Id: Ia87e8a64fc47ceb990283798cbd7654897222fb6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76483
Tested-by: Frank Nägler <frank.naegler@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>