[TASK] Move record list clipboard and searchbox toggles to view settings
With this patch, we are starting to declutter the record list.
The search box and Clipboard toggles are now available in the "View"
settings of the record list view. This change follows the pattern we
already introduced in the file list. For the search box, the level
selector is now always visible.

Resolves: #99895
Releases: main
Change-Id: I55cd65f3eda491d44199d9886610c4d6135acc38
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77793
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>