[BUGFIX] Handle unavailability of (dis)allowed extensions in FileBrowser
For cases when the FileBrowser doesn't receive any information
of allowed or disallowed file extensions via the "bparams" query
parameter, the code is slightly adjusted to handle this case
to avoid PHP errors.

Additonally, the avatar field of the user settings is adjusted
to provide corresponding information, while the allowed types
are set to the "imagefile_ext" list. Previously used "allowed"
option in the configuration of the field is removed as it had
no effect. The avatar image is always checked against the
"imagefile_ext" list.

Resolves: #101342
Related: #101231
Releases: main, 12.4
Change-Id: I9302315110d418e277e733a8a767de595b611efb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80017
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan B�rk <stefan@buerk.tech>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>