[!!!][FEATURE] Add UriBuilder->buildUriFromRequest
This change adds a new method buildUriFromRequest
inside TYPO3's Backend UriBuilder in order to
generate links to the current (or any) request
with additional parameters.

This is quite helpful to link to the same module
/route when a request is given, thus reducing
code duplication and unifies link generation.

In addition, the determineScriptUrl() and
getScriptUrl() methods inside LinkBrowsers
and ElementBrowsers are removed, and have been
removed from the LinkParameterProviderInterface
as well.

Resolves: #101612
Releases: main
Change-Id: Ibaa2d2872178374e4553cf5899ee8ae19168480b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80433
Reviewed-by: Stefan B�rk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan B�rk <stefan@buerk.tech>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>