[BUGFIX] Restore currently selected folder in ElementBrowser
Since the major refactoring of ElementBrowsers into the file list module
the order in which the currently selected folder from the user is loaded
got messed up. The retrieval of the folder was moved from the
render method to the initialization chain. However, at this point the
session data is not processed yet.

To fix this, the folder retrieval is moved back to the start of the
render method.

Resolves: #100837
Related: #99861
Releases: main, 12.4
Change-Id: I410306c8c1641e538367d6773474a660e9f383e8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80573
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>