[BUGFIX] Guard local storage with missing base folder in filelist
Storages using the local storage driver with a non existing
base folder are put offline automatically. However, the file
list module displays a to generic error message not pointing
to that fact.

This change adds a concrete check for this case and displays
a more specific error flash message in the file list module now.

This can be tested by creating a additional local storage with
a non-existing relative path, for example:

    `fileadmin/not-existing-folder/`

Resolves: #85323
Releases: main, 12.4, 11.5
Change-Id: Ie3ad64a99c0871d258470ffc858639dee4de1108
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81277
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>