[BUGFIX] Omit click menu on icons in `browse` mode
The element browsers, such as file browser,
folder browser or the create folder browser
are rendering the FileList in the `browse`
mode. In this case the context menu component
is not loaded. This is correct, since using
the context menu inside the modal would lead
to various side effects.

Because the JS component is not loaded for
mentioned reasons, wrapping the icons with
the click menu needs to be omitted. This
otherwise leads to further side effect,
e.g. loading a second folder tree in the
create folder browser.

Therefore, this change now omits wrapping
the icon with then click menu when the
Filelist is used in the `browse` mode.

Resolves: #103001
Releases: main, 12.4
Change-Id: Ifd4daf5487b5dd9c74553d48cd0c2270b2c4f58b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82742
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>