[TASK] Add meaningful title to File selector add file button
The "+" labelled button in the File selector modal was just
labelled with the filename as the title which doesn't give
any meaning to screenreader users.

To show the purpose of the button to screenreader users the
title has been changed to additionally describe the action.

Also an aria-label with the same value has been added. This
is not ideal, because some screenreaders will read out the
title twice, but on the other hand other screenreaders ignore
the title or the aria-label attribute.

Resolves: #103292
Releases: main, 12.4
Change-Id: Ia53fd77623e3cbd3a63a48f1bd4edc457f453c6b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83224
Tested-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>