[BUGFIX] Check language access for edit default metadata
At various places, it's possible to edit the metadata
of a file in the default language. However, users might
not have access to the default language. To prevent
moving to FormEngine and facing access permission
errors, those places now use proper language access
checks and do not display corresponding action if
the user does not have access.

Resolves: #103432
Releases: main, 12.4
Change-Id: I22c3755a17888cfc623ce05b25fd655b323cc553
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83583
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>