[BUGFIX] Properly respect "enableClipBoard" TSconfig
With the rework of moving the "Show clipboard" toggle into the view
mode dropdown for the list and filelist module, the functionality
of the user TSconfig "options.file_list.enableClipBoard" and
the page TSconfig "mod.web_list.enableClipBoard" has been
broken. This commit fixes the functionality, aligning it with the
documented behaviours.

Resolves: #102145
Releases: main, 12.4
Change-Id: I17e3677638ea9d2f950a9f8d6c33ad089fc813c1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84666
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>