[BUGFIX] Store selected folder in link browser
Since TYPO3 version 12.3, the currently selected folder was
no longer stored in the file link browser after refactoring.

The previous behaviour is now restored: If the link browser
is opended multiple times or the user switches to the file
tab again, the last selected folder is selected again by
storing this information for the current user.

Resolves: #104603
Related: #100206
Releases: main, 13.4, 12.4
Change-Id: Ie1dc851e4b967ca941746a246447ebb146c8c41c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87301
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>