[BUGFIX] Add margin to upload button in filelist module
When an empty folder is displayed in the filelist module,
an upload button is shown. Clicking this button leads to a
dedicated upload view. In this view, an infobox is shown,
which has no top margin to the previous upload button.

This change adds the necessary margin to the upload
button div.

Resolves: #105862
Releases: main, 13.4, 12.4
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: Ib1e779edb7476639a262a4198fe42b70b0fa1d58
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87592
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>