[BUGFIX] Update FAL error handling in link browser
Within file and folder link browsers, the latest selected
folder a user selected is stored in session data, and is
tried to be selected again on next use.

If a folder or an entire storage has been removed meanwhile,
or permissions changed and the user no longer has access,
hydration of the selected folder fails, which is not
properly sanitized.

The patch simplifies the costruct to omit guessing a
different folder that is marked as auto selected in such
cases and sanitizes the FAL API calls to avoid exceptions.

Resolves: #105152
Releases: main, 13.4, 12.4
Change-Id: Idd52423cefe4709fd3b8dfce233fbfd842840ad3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87957
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>