[TASK] Avoid loading metadata for files when obtaining the "file" property
Because FAL loads all files in one folder before applying a pagination
strategy, custom drivers, that make use of heavy i/o tasks may have very
poor performance, especially when dealing with many files in one folder.

By avoiding access to the metadata of a file the response time can still be
reduced substantially.

The default file-list rendering will sort by "file" which is the sys_file
uid but will be obtained via the metadata aspect. This is not necessary as
the file object should be able to return the "name" without using the
metadata aspect.

Releases: main, 13.4, 12.4
Resolves: #106248
Change-Id: I286f39d57f9f9e5a192a3148ba9953e5abf8f26d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88355
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>