[!!!][TASK] Adapt FAL codebase - Part 2: Align PHP Interfaces
In order to coordinate FAL code, this change
adapts FileInterface and FolderInterface as well as AbstractFile

- AbstractFile->rename()
- AbstractFile->copyTo()
- AbstractFile->moveTo()

are altogether moved to the File object, as this should not
be exposed to e.g. ProcessedFiles.

For this reason "FileInterface::rename()" is removed from
the FileInterface.

FolderInterface has more methods to implement now,
in order to allow interchangeability:
- FolderInterface::getSubFolder() returns a FolderInterface instance
- FolderInterface::getReadablePath() is added
- FolderInterface::getFiles() is added
- FolderInterface::searchFiles() is added
- FolderInterface::getRole() is added

Some calling code is adapted thanks to PHPstan.

Once this cleanup is done, more work on ResourceStorage is needed
to handle differentiation between File|ProcessedFile|null
as in getFileByIdentifier().

Resolves: #106524
Related: #106427
Releases: main
Change-Id: Id0ec8aa25278e4166d432b56547047fe4bdfe0a4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88949
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Georg Ringer <georg.ringer@gmail.com>