[BUGFIX] Always sort items "naturally" in filelist
The collator option `\Collator::NUMERIC_COLLATION` is used to sort its
items in a natural manner as understood by human beings, which is now
always enabled.  Previously, the flag was only used when sorting by file
size.

Additionally, when reversing the result is requested, the calculation is
adjusted to handle this directly in `uksort()`, making the call of
`array_reverse()` obsolete.

Due to the removal of `array_reverse()`, array keys are not reset
anymore, therefore the tests have to use `assertEqualsCanonicalizing()`
as we only care about the correct order of the values. To have _this_
functional, we may not use mocked `ResourceInterface` instances due to
recursion limits.

Resolves: #106607
Releases: main, 13.4
Change-Id: I95cbe3fc3b99de6bbd849e7ffca776c74fcd3449
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89239
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>