[BUGFIX] Add global sorting dropdown for files
In order to allow sorting files in tiles view, a consolidated sorting
dropdown button is introduced in the file list which is active in both
view modes.
The same sorting dropdown is applied to instances of Element Browser and
Link Browser as well.

To represent the sorting direction, a new enum `SortDirection` is
introduced for EXT:filelist and incorporated into existing code. The
render mode enum `Enum` is now able to return an array containing the
fields the file list can be ordered by.

The default sorting by "file" is changed to "name" as the column
"file" does not exist. Also, sorting by "RW" has been disabled as it
actively doesn't fulfill any purpose.

Resolves: #105952
Releases: main, 13.4
Change-Id: Ib7631e23ce6bdc69d4257090d916d83d78ec815a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89786
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>