[BUGFIX] Handle oversights in thumbnail web component
The method `ThumbnailSize::getDimensions()` falsely used plain integer
values when rendering thumbnails in case aspect ratio should be kept.
This caused distorted images as the rendered thumbnail was forced in the
configured dimensions (e.g. 64x64). This commit fixes this by using the
`m` suffix, as documented at [1].

The `width` and `height` attributes should be applied to the rendered
image only if they are passed with a reasonable value. If the arguments
are omitted, they get the value `0` assigned, leading to broken
rendering. The code now checks whether their respective value is larger
than 0 before applying them to the image.

The need to create the URL to the `ResourceController` via PHP is not
necessary, as this implies code duplication when called at other places.
For this reason, the "base URL" is generated on a global level and
passed down as `TYPO3.settings.Resource.thumbnailUrl`.

The `ThumbnailSize` enum was falsely stored in EXT:filelist, creating
an unwanted dependency within EXT:backend. The enum has been moved to a
more appropriate location.

[1] https://docs.typo3.org/m/typo3/reference-typoscript/main/en-us/Functions/Imgresource.html#width

Resolves: #107162
Related: #106902
Releases: main, 13.4, 12.4
Change-Id: I0a20cf2310b19a72254af3b1a535fb71a70d98a8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90184
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>