[TASK] Remove justLocalized parameter and use localization wizard
The legacy "justLocalized" parameter was used to redirect to
the edit form of a newly created translated record. This
approach required a two-step redirect process:
1. Redirect to tce_db to create the localization
2. Redirect back to record_edit with justLocalized parameter
3. localizationRedirect() looks up the translated record
4. Final redirect to edit the translated record

This patch replaces this mechanism with the modern localization
wizard approach which uses the new
typo3-backend-localization-button web component

This provides a more consistent translation workflow across
the backend and eliminates the need for complex redirect
chains.

Resolves: #108512
Releases: main
Change-Id: Iac8f66c852bb4aa8f9bcdbebf24ab7b9282b896e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91663
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: core-ci <typo3@b13.com>