[TASK] Use native ProgressBarElement instead of nprogress
In order to use our custom styling and rendering for
the new Backend UI, the nprogress loading bar
is replaced by our custom <typo3-backend-progress-bar>
web component.

This not only removes a third-party library with
a more modern approach, it also allows to customize
the color of the progress bar based on the themes' primary
color.

FYI: The last nprogress release was in 2015.

Resolves: #108589
Releases: main
Change-Id: Ie81942837e6949a0993eb1fff1986b4d054d9752
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/92149
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Jochen Roth <rothjochen@gmail.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>